backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow==9.5.0
psycopg2-binary==2.9.9
python-decouple==3.8
opencv-python==4.8.1.78
numpy==1.24.3
tensorflow==2.13.0
easyocr==1.7.0
spacy==3.7.2
python-dateutil==2.8.2
celery==5.3.4
redis==5.0.1
django-storages==1.14.2
boto3==1.34.0
fuzzywuzzy==0.18.0
python-levenshtein==0.25.1
langdetect==1.0.9
paddlepaddle>=2.6.0
paddleocr>=2.7.0
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0
scikit-learn==1.3.2
datasets==2.14.6
sentencepiece==0.1.99
sentencepiece==0.1.99

